R6MD 1.1.0: team-index mapping, Overwolf bridge, live layer, replay analytics

+++ src/R6Md/Replays/ReplayPayload.cs
@@ -69,7 +69,7 @@ public sealed record ReplayPlayerStat(
 	[property: JsonPropertyName("died")] bool? Died,
 	[property: JsonPropertyName("assists")] int? Assists,
 	[property: JsonPropertyName("headshots")] int? Headshots,
-	[property: JsonPropertyName("headshotPercentage")] int? HeadshotPercentage,
+	[property: JsonPropertyName("headshotPercentage")] double? HeadshotPercentage,
 	[property: JsonPropertyName("damageTaken")] int? DamageTaken,
 	[property: JsonPropertyName("damageDealt")] int? DamageDealt,
 	[property: JsonPropertyName("secondsAlive")] int? SecondsAlive);
